+++ Sources/KeyboardKit/KeyboardInputViewController.swift
@@ -106,10 +106,8 @@ open class KeyboardInputViewController: UIInputViewController {
      This stays the same even if a ``textInputProxy`` is set,
      which makes ``textDocumentProxy`` return the custom one
      instead of the original one.
-
-     TODO: Rename this to `mainTextDocumentProxy`.
      */
-    open var originalTextDocumentProxy: UITextDocumentProxy {
+    open var mainTextDocumentProxy: UITextDocumentProxy {
         super.textDocumentProxy
     }
 
@@ -119,7 +117,7 @@ open class KeyboardInputViewController: UIInputViewController {
      it is set to a custom value.
      */
     open override var textDocumentProxy: UITextDocumentProxy {
-        textInputProxy ?? originalTextDocumentProxy
+        textInputProxy ?? mainTextDocumentProxy
     }
     
     /**
@@ -127,8 +125,6 @@ open class KeyboardInputViewController: UIInputViewController {
 
      Setting the property makes ``textDocumentProxy`` return
      the custom proxy instead of the original one.
-
-     TODO: Rename this to `keyboardInputProxy`.
      */
     public var textInputProxy: TextInputProxy? {
         didSet { viewWillSyncWithContext() }