+++ RELEASE_NOTES.md
@@ -96,9 +96,11 @@ If you have problems upgrading to `7.0`, first try upgrading to `6.9`. It has a 
 * `KeyboardAction` `.shift` `currentState` is renamed to `currentCasing`.
 * `KeyboardAction` `standardTextDocumentProxyAction` has been removed.
 * `KeyboardAppearance` callout styles have been converted to properties.
+* `KeyboardContext` `originalTextDocumentProxy` has been renamed to `mainTextDocumentProxy`.
 * `KeyboardGesture` `.tap` has been replaced by `.release`.
 * `KeyboardInputViewController` has replaced `actionCalloutContext` and `inputCalloutContext` with a single `calloutContext`.
 * `KeyboardInputViewController` `.shared` has been removed.
+* `KeyboardInputViewController` `originalTextDocumentProxy` has been renamed to `mainTextDocumentProxy`.
 * `KeyboardReturnActionMappable` has been refactored to `KeyboardActionMappable`.
 * `KeyboardTextField` now requires an keyboard input view controller.
 * `KeyboardTextView` now requires an keyboard input view controller.